# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Show Download Manager API",
        "title": "Gongon",
        "contact": {
            "name": "Jusoares",
            "email": "julianosgreg@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/database/indexer": {
            "get": {
                "description": "List all added indexers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database/Indexer"
                ],
                "summary": "List Indexers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add Indexer to db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database/Indexer"
                ],
                "summary": "Add Indexer",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.IdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Indexer from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database/Indexer"
                ],
                "summary": "Delete Indexer",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.IdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/database/show/episode/status": {
            "post": {
                "description": "Change Episode Tracking Status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database/Show/Episode"
                ],
                "summary": "Change Episode Status",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/episode.ChangeEpisodeTrackingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/database/show/episode/{id}": {
            "get": {
                "description": "List Show Episodes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database/Episodes"
                ],
                "summary": "List Show Episodes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Show ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/prowlarr/search": {
            "post": {
                "description": "Search shows on prowlarr indexers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prowlarr/Search"
                ],
                "summary": "Search For Shows",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.SearchResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tvmaze/search/name": {
            "post": {
                "description": "Search TvMaze By Show Name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TvMaze/Search"
                ],
                "summary": "Seach Show By Name",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.NameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tvmaze/search/tvmaze": {
            "post": {
                "description": "Search TvMaze By TvMaze ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TvMaze/Search"
                ],
                "summary": "Seach Show By TvMaze ID",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.IdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "episode.ChangeEpisodeTrackingRequest": {
            "type": "object",
            "properties": {
                "episode_id": {
                    "type": "integer"
                },
                "tracking": {
                    "type": "string"
                }
            }
        },
...
        "schema.SearchRequest": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "schema.SearchResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "number"
                },
                "ageHours": {
                    "type": "number"
                },
                "ageMinutes": {
                    "type": "number"
                },
                "fileName": {
                    "type": "string"
                },
                "grabs": {
                    "type": "integer"
                },
                "guid": {
                    "type": "string"
                },
                "indexer": {
                    "type": "string"
                },
                "indexerId": {
                    "type": "integer"
                },
                "infoHash": {
                    "type": "string"
                },
                "infoUrl": {
                    "type": "string"
                },
                "leechers": {
                    "type": "integer"
                },
                "magnetUrl": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "publishDate": {
                    "type": "string"
                },
                "seeders": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schemas.DefaultResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "schemas.IdRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "schemas.NameRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}